+++ apps/demo/CHANGELOG.md
@@ -1,5 +1,26 @@
 # demo
 
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @tanggoat/request@1.0.5
+
+## 0.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @tanggoat/request@1.0.4
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @tanggoat/request@1.0.3
+
 ## 0.0.1-beta.0
 
 ### Patch Changes